Continuing to work on finalizing code to submit for revision #2 to Bioinformatics. Mainly removed a lot of unnecessary legacy code/functions. More specifically:
CallMatlab: Removed SplitTensor R function, as it is not used.

DataProc: Removed RemoveMissingDataFromXY, as it is not used.

EvaluateTensor: Removed ComputeErrorRate, SummarizeOutput, as it is not used.

Stripped GSEA src code into the only functions used in this package, namely RunGsea, GetGSEAResultsV2, and AppendGSTerms.

Removed test_helpers.R - nothing in there was being used!